Refactoring, separated frontend and backend
