+++ src/frontend/components/Home.js
@@ -3,7 +3,7 @@ import Header from './Header';
 import Body from './Body';
 
 
-export default class Home extends React.Component {
+class Home extends React.Component {
   render() {
     return (
       <div>
@@ -13,3 +13,5 @@ export default class Home extends React.Component {
     );
   }
 }
+
+export default Home;
