+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/App';
-import * as serviceWorker from './components/serviceWorker';
+import App from './frontend/App';
+import * as serviceWorker from './frontend/serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
